# Conflicts:
#	src/main/generated/org/webMonster/uniManageBoot/member/entity/QMemberEntity.java
#	src/main/generated/org/webMonster/uniManageBoot/professor/cancelledLecture/entity/QCancelledLectureEntity.java
#	src/main/generated/org/webMonster/uniManageBoot/professor/lecture/entity/QLectureEntity.java
#	src/main/generated/org/webMonster/uniManageBoot/professor/lectureClass/entity/QLectureClassEntity.java
#	src/main/generated/org/webMonster/uniManageBoot/student/courseRegi/entity/QCourseRegiEntity.java
#	src/main/java/org/webMonster/uniManageBoot/professor/lecture/model/service/LectureService.java
#	src/main/resources/static/js/about.js

+++ src/main/java/org/webMonster/uniManageBoot/member/model/dto/MemberDto.java
@@ -18,7 +18,7 @@ public class MemberDto {
     private long memberId;  //아이디
     private String memberPwd;  //비밀번호
     private String name;        //이름
-    private int departmentId;  //학과
+    private long departmentId;  //학과
     private int grade;        //학년
     private Date birthday;      //생년월일
     private String phone;       //전화번호
